improve file structure
