+++ ddtrace/llmobs/experimentation/__init__.py
@@ -2,11 +2,11 @@
 LLM Datasets and Experiments.
 """
 
-from ._experiments import Dataset
-from ._experiments import Experiment
-from ._experiments import task
-from ._experiments import evaluator
-from ._experiments import init
+from ._dataset import Dataset
+from ._experiment import Experiment
+from ._decorators import task
+from ._decorators import evaluator
+from ._config import init
 
 
 __all__ = ["Dataset", "Experiment", "task", "evaluator", "init"]
