+++ ddtrace/llmobs/experimentation/_experiment.py
@@ -1,690 +1,21 @@
 import concurrent.futures
 from copy import deepcopy
-import csv
-from enum import Enum
-from functools import wraps
-import inspect
 import json
 import os
 import time
 import traceback
-from typing import Any, Callable, Dict, Iterator, List, Optional, Union
-from urllib.parse import quote
-import uuid
+from typing import Any, Callable, Dict, List, Optional, Iterator, TYPE_CHECKING
 
 import ddtrace
 
+from ._dataset import Dataset
+from .utils._http import exp_http_request
+from ._config import ENV_PROJECT_NAME, _validate_init, _is_locally_initialized, DEFAULT_CONCURRENT_JOBS, BASE_URL, DEFAULT_CHUNK_SIZE
+from .utils._ui import Color, ProgressReporter, _print_progress_bar
 from .._llmobs import LLMObs
-from .._utils import HTTPResponse, http_request
 
-
-# Default configuration values
-DEFAULT_SITE = "datadoghq.com"
-DEFAULT_ML_APP = "dne"
-MAX_DATASET_ROWS = 50000
-MAX_PROGRESS_BAR_WIDTH = 40
-DEFAULT_CHUNK_SIZE = 300
-DEFAULT_CONCURRENT_JOBS = 10
-
-# Global state (initialized by init())
-IS_INITIALIZED = False
-ENV_PROJECT_NAME = None
-ENV_DD_SITE = DEFAULT_SITE
-ENV_DD_API_KEY = None
-ENV_DD_APPLICATION_KEY = None
-ML_APP = DEFAULT_ML_APP
-RUN_LOCALLY = False
-
-# Derived values
-BASE_URL = f"https://api.{ENV_DD_SITE}"
-
-
-def init(
-    project_name: str,
-    site: str = DEFAULT_SITE,
-    api_key: str = None,
-    application_key: str = None,
-    run_locally: bool = False,
-) -> None:
-    """
-    Initialize an experiment environment for pushing and retrieving data from Datadog.
-
-    This function configures global settings needed to interact with the Datadog LLM Observability API.
-    If 'run_locally' is True, no data will be sent; otherwise, this function ensures that all required
-    credentials are available. If no credentials are passed in, it attempts to read them from environment
-    variables.
-
-    Args:
-        project_name (str): Name of the project to be associated with any experiments.
-        site (str, optional): Datadog site domain (e.g., "datadoghq.com"). Defaults to "datadoghq.com".
-        api_key (str, optional): Datadog API key. Reads from environment variable DD_API_KEY if None.
-        application_key (str, optional): Datadog Application key. Reads from environment variable DD_APPLICATION_KEY if None.
-        run_locally (bool, optional): If True, disables communication with Datadog and runs locally.
-
-    Raises:
-        ValueError: If required environment variables or parameters are missing when run_locally is False.
-    """
-
-    global IS_INITIALIZED, ENV_PROJECT_NAME, ENV_DD_SITE, ENV_DD_API_KEY, ENV_DD_APPLICATION_KEY, RUN_LOCALLY
-
-    if run_locally:
-        RUN_LOCALLY = True
-
-        return
-
-    if api_key is None:
-        api_key = os.getenv("DD_API_KEY")
-        if api_key is None:
-            raise ValueError(
-                "DD_API_KEY environment variable is not set, please set it or pass it as an argument to init(...api_key=...)"
-            )
-
-    if application_key is None:
-        application_key = os.getenv("DD_APPLICATION_KEY")
-        if application_key is None:
-            raise ValueError(
-                "DD_APPLICATION_KEY environment variable is not set, please set it or pass it as an argument to init(...application_key=...)"
-            )
-
-    if site is None:
-        site = os.getenv("DD_SITE")
-        if site is None:
-            raise ValueError(
-                "DD_SITE environment variable is not set, please set it or pass it as an argument to init(...site=...)"
-            )
-
-    if IS_INITIALIZED is False:
-        LLMObs.enable(
-            ml_app=ML_APP,
-            integrations_enabled=True,
-            agentless_enabled=True,
-            site=site,
-            api_key=api_key,
-        )
-
-    ENV_PROJECT_NAME = project_name
-    ENV_DD_SITE = site
-    ENV_DD_API_KEY = api_key
-    ENV_DD_APPLICATION_KEY = application_key
-    IS_INITIALIZED = True
-
-
-def _validate_init() -> None:
-    """
-    Check if the environment has been initialized.
-
-    Raises:
-        ValueError: If the environment is not yet initialized (init() has not been called).
-    """
-    if IS_INITIALIZED is False:
-        raise ValueError(
-            "Environment not initialized, please call ddtrace.llmobs.experiments.init() at the top of your script before calling any other functions"
-        )
-
-
-def _is_locally_initialized() -> bool:
-    """
-    Check if the environment is configured to run locally.
-
-    Returns:
-        bool: True if running locally, False otherwise.
-    """
-    return RUN_LOCALLY
-
-
-class Dataset:
-    """
-    A container for LLM experiment data that can be pushed to and retrieved from Datadog.
-
-    This class manages collections of input-output pairs used for LLM experiments,
-    providing functionality to validate data, pull or push datasets to Datadog, and
-    retrieve them as Pandas DataFrames.
-
-    Attributes:
-        name (str): Name of the dataset.
-        description (str): Optional description of the dataset.
-        version (int): Version of the dataset (not necessarily the same as the Datadog dataset version).
-    """
-
-    def __init__(
-        self, name: str, data: Optional[List[Dict[str, Union[str, Dict[str, Any]]]]] = None, description: str = ""
-    ) -> None:
-        """
-        Initialize a Dataset instance, optionally loading data from Datadog if none is provided.
-
-        Args:
-            name (str): Name of the dataset.
-            data (List[Dict[str, Union[str, Dict[str, Any]]]], optional): List of records where each record
-                must contain 'input' and 'expected_output' fields. Both values can be strings or dictionaries.
-                If None, attempts to pull the dataset from Datadog.
-            description (str, optional): Optional description of the dataset. Defaults to "".
-
-        Raises:
-            ValueError: If the data is invalid (violates structure or size constraints).
-        """
-        self.name = name
-        self.description = description
-        self.version = 0
-
-        # If no data provided, attempt to pull from Datadog
-        if data is None:
-            print(f"No data provided, pulling dataset '{name}' from Datadog...")
-            pulled_dataset = self.pull(name)
-            print(pulled_dataset._datadog_dataset_id)
-            self._data = pulled_dataset._data
-            self._datadog_dataset_id = pulled_dataset._datadog_dataset_id
-            self._version = pulled_dataset._datadog_dataset_version
-        else:
-            self._validate_data(data)
-            self._data = data
-            self._datadog_dataset_id = None
-            self._version = 0
-
-    def __iter__(self) -> Iterator[Dict[str, Union[str, Dict[str, Any]]]]:
-        """
-        Create an iterator over the dataset records.
-
-        Returns:
-            Iterator[Dict[str, Union[str, Dict[str, Any]]]]: An iterator of dataset records.
-        """
-        return iter(self._data)
-
-    def __len__(self) -> int:
-        """
-        Return the number of records in the dataset.
-
-        Returns:
-            int: The size of the dataset.
-        """
-        return len(self._data)
-
-    def __getitem__(self, index: int) -> Dict[str, Union[str, Dict[str, Any]]]:
-        """
-        Retrieve a specific dataset record by index.
-
-        Args:
-            index (int): Index of the record to retrieve.
-
-        Returns:
-            Dict[str, Union[str, Dict[str, Any]]]: A copy of the dataset record at the given index.
-        """
-        record = self._data[index].copy()
-        return record
-
-    def _validate_data(self, data: List[Dict[str, Union[str, Dict[str, Any]]]]) -> None:
-        """
-        Validate the structure and size of dataset records.
-
-        Args:
-            data (List[Dict[str, Union[str, Dict[str, Any]]]]): List of dataset records.
-
-        Raises:
-            ValueError: If the data is empty, exceeds 50,000 records, or records are inconsistent in keys or data type.
-        """
-        if not data:
-            raise ValueError("Data cannot be empty.")
-
-        if len(data) > MAX_DATASET_ROWS:
-            raise ValueError(f"Dataset cannot exceed {MAX_DATASET_ROWS} rows.")
-
-        if not all(isinstance(row, dict) for row in data):
-            raise ValueError("All rows must be dictionaries.")
-
-        first_row_keys = set(data[0].keys())
-        for row in data:
-            if set(row.keys()) != first_row_keys:
-                raise ValueError("All rows must have the same keys.")
-
-    @classmethod
-    def pull(cls, name: str, version: int = None) -> "Dataset":
-        """
-        Create a dataset from an existing dataset hosted in Datadog.
-
-        Args:
-            name (str): Name of the dataset to retrieve from Datadog.
-            version (int, optional): Specific version of the dataset to retrieve. If None, uses the latest version.
-
-        Returns:
-            Dataset: A new Dataset instance populated with the records from Datadog.
-
-        Raises:
-            ValueError: If the dataset is not found or if the specified version doesn't exist.
-            Exception: If any HTTP or unexpected error occurs during the request.
-        """
-        _validate_init()
-        # Get dataset ID
-        encoded_name = quote(name)
-        url = f"/api/unstable/llm-obs/v1/datasets?filter[name]={encoded_name}"
-
-        resp = exp_http_request("GET", url)
-        response_data = resp.json()
-        datasets = response_data.get("data", [])
-
-        if not datasets:
-            raise ValueError(f"Dataset '{name}' not found")
-
-        dataset_id = datasets[0]["id"]
-        dataset_version = datasets[0]["attributes"]["current_version"]
-
-        print(f"Found dataset '{name}' (latest version: {dataset_version})")
-
-        # If version specified, verify it exists
-        if version is not None:
-            versions_url = f"/api/unstable/llm-obs/v1/datasets/{dataset_id}/versions"
-            try:
-                versions_resp = exp_http_request("GET", versions_url)
-                versions_data = versions_resp.json()
-                available_versions = [v["attributes"]["version_number"] for v in versions_data.get("data", [])]
-
-                if not available_versions:
-                    raise ValueError(f"No versions found for dataset '{name}'")
-
-                if version not in available_versions:
-                    versions_str = ", ".join(str(v) for v in sorted(available_versions))
-                    raise ValueError(
-                        f"Version {version} not found for dataset '{name}'. " f"Available versions: {versions_str}"
-                    )
-                dataset_version = version
-            except ValueError as e:
-                if "404" in str(e):
-                    raise ValueError(f"Dataset '{name}' not found") from e
-                raise
-
-        # Get dataset records
-        url = f"/api/unstable/llm-obs/v1/datasets/{dataset_id}/records"
-        if version is not None:
-            url += f"?filter[version]={version}"
-
-        try:
-            resp = exp_http_request("GET", url)
-            records_data = resp.json()
-        except ValueError as e:
-            if "404" in str(e):
-                if version is not None:
-                    versions_str = ", ".join(str(v) for v in sorted(available_versions))
-                    raise ValueError(
-                        f"Version {version} not found for dataset '{name}'. " f"Available versions: {versions_str}"
-                    ) from e
-                raise ValueError(f"Dataset '{name}' not found") from e
-            raise
-
-        if not records_data.get("data", []):
-            raise ValueError(f"Dataset '{name}' does not contain any records.")
-
-        # Transform records into the expected format
-        class_records = []
-        for record in records_data.get("data", []):
-            attrs = record.get("attributes", {})
-            input_data = attrs.get("input")
-            expected_output = attrs.get("expected_output")
-
-            class_records.append(
-                {
-                    "record_id": record.get("id"),
-                    "input": input_data,
-                    "expected_output": expected_output,
-                    **attrs.get("metadata", {}),
-                }
-            )
-
-        # Create new dataset instance
-        dataset = cls(name, class_records)
-        dataset._datadog_dataset_id = dataset_id
-        dataset._datadog_dataset_version = dataset_version
-        return dataset
-
-    def push(self, overwrite: bool = False) -> None:
-        """
-        Push the dataset to Datadog, optionally overwriting an existing dataset.
-
-        This method either updates an existing dataset (if found and 'overwrite' is True)
-        or creates a new dataset. If the dataset is newly created, a new version is also created
-        in Datadog. After pushing, the local dataset is refreshed with the latest information
-        from Datadog, including record IDs and metadata.
-
-        Args:
-            overwrite (bool, optional): If True, overwrite the existing dataset if found instead
-                of creating a new version. Defaults to False.
-
-        Raises:
-            ValueError: If environment is not initialized or any other HTTP error occurs during push.
-        """
-        _validate_init()
-
-        # Reasonable chunk size for batching requests
-        chunk_size: int = DEFAULT_CHUNK_SIZE
-
-        # First check if dataset exists
-        encoded_name = quote(self.name)
-        url = f"/api/unstable/llm-obs/v1/datasets?filter[name]={encoded_name}"
-        resp = exp_http_request("GET", url)
-        existing_dataset = resp.json().get("data", [])
-
-        if existing_dataset and overwrite:
-            # Use existing dataset ID
-            dataset_id = existing_dataset[0]["id"]
-            self._datadog_dataset_id = dataset_id
-            self._datadog_dataset_version = existing_dataset[0]["attributes"]["current_version"]
-
-            # Get existing records to map record IDs
-            url = f"/api/unstable/llm-obs/v1/datasets/{dataset_id}/records"
-            resp = exp_http_request("GET", url)
-            existing_records = resp.json().get("data", [])
-
-            # Update records in chunks
-            total_records = len(self._data)
-            chunks = [self._data[i : i + chunk_size] for i in range(0, total_records, chunk_size)]
-            total_chunks = len(chunks)
-
-            show_progress = total_records > chunk_size
-            if show_progress:
-                _print_progress_bar(0, total_chunks, prefix="Updating:", suffix="Complete")
-
-            for i, chunk in enumerate(chunks):
-                for record, existing in zip(chunk, existing_records[i * chunk_size : (i + 1) * chunk_size]):
-                    record_id = existing["id"]
-                    payload = {
-                        "data": {
-                            "type": "datasets",
-                            "attributes": {
-                                "input": record["input"],
-                                "expected_output": record["expected_output"],
-                                "metadata": {
-                                    k: v
-                                    for k, v in record.items()
-                                    if k not in ["input", "expected_output", "record_id"]
-                                },
-                            },
-                        }
-                    }
-                    url = f"/api/unstable/llm-obs/v1/datasets/{dataset_id}/records/{record_id}"
-                    resp = exp_http_request("PATCH", url, body=json.dumps(payload).encode("utf-8"))
-
-                if show_progress:
-                    _print_progress_bar(i + 1, total_chunks, prefix="Updating:", suffix="Complete")
-
-        else:
-            # Create new dataset
-            dataset_payload = {
-                "data": {
-                    "type": "datasets",
-                    "attributes": {
-                        "name": self.name,
-                        "description": self.description,
-                    },
-                }
-            }
-            resp = exp_http_request(
-                "POST", "/api/unstable/llm-obs/v1/datasets", body=json.dumps(dataset_payload).encode("utf-8")
-            )
-            response_data = resp.json()
-            dataset_id = response_data["data"]["id"]
-            self._datadog_dataset_id = dataset_id
-            self._datadog_dataset_version = 0
-
-            # Split records into chunks and upload
-            total_records = len(self._data)
-            chunks = [self._data[i : i + chunk_size] for i in range(0, total_records, chunk_size)]
-            total_chunks = len(chunks)
-
-            # Only show progress bar for large datasets
-            show_progress = total_records > chunk_size
-            if show_progress:
-                _print_progress_bar(0, total_chunks, prefix="Uploading:", suffix="Complete")
-
-            for i, chunk in enumerate(chunks):
-                records_payload = {"data": {"type": "datasets", "attributes": {"records": chunk}}}
-                url = f"/api/unstable/llm-obs/v1/datasets/{dataset_id}/records"
-                resp = exp_http_request("POST", url, body=json.dumps(records_payload).encode("utf-8"))
-
-                if show_progress:
-                    _print_progress_bar(i + 1, total_chunks, prefix="Uploading:", suffix="Complete")
-
-            time.sleep(1)  # Sleep to allow for processing after ingestion.
-            # Pull the dataset to get all record IDs and metadata
-            pulled_dataset = self.pull(self.name)
-            self._data = pulled_dataset._data
-            self._datadog_dataset_id = pulled_dataset._datadog_dataset_id
-            self._datadog_dataset_version = pulled_dataset._datadog_dataset_version
-
-            # Print url to the dataset in Datadog
-            print(f"\nDataset '{self.name}' created: {BASE_URL}/llm/testing/datasets/{dataset_id}\n")
-
-    @classmethod
-    def from_csv(
-        cls,
-        filepath: str,
-        name: str,
-        description: str = "",
-        delimiter: str = ",",
-        input_columns: List[str] = None,
-        expected_output_columns: List[str] = None,
-    ) -> "Dataset":
-        """
-        Create a Dataset from a CSV file by specifying which columns correspond
-        to inputs and which columns correspond to expected outputs.
-
-        Args:
-            filepath (str): Path to the CSV file.
-            name (str): Name of the dataset (this will be used in Datadog).
-            description (str, optional): Optional description of the dataset. Defaults to "".
-            delimiter (str, optional): CSV delimiter character, defaults to comma.
-            input_columns (List[str], optional): List of column names that should be treated as input data.
-            expected_output_columns (List[str], optional): List of column names that should be treated as expected output data.
-
-        Returns:
-            Dataset: A new Dataset instance containing the CSV data, structured for LLM experiments.
-
-        Raises:
-            ValueError: If input_columns or expected_output_columns are not provided,
-                or if the CSV is missing those columns, or if the file is empty.
-            DatasetFileError: If there are issues reading the CSV file (e.g., file not found or IO permission error).
-        """
-        if input_columns is None or expected_output_columns is None:
-            raise ValueError("`input_columns` and `expected_output_columns` must be provided.")
-
-        data = []
-        try:
-            with open(filepath, mode="r", encoding="utf-8") as csvfile:
-                reader = csv.DictReader(csvfile, delimiter=delimiter)
-                rows = list(reader)
-                if not rows:
-                    raise ValueError("CSV file is empty.")
-
-                # Ensure that the specified columns are present
-                header_columns = reader.fieldnames
-                missing_input_columns = [col for col in input_columns if col not in header_columns]
-                missing_output_columns = [col for col in expected_output_columns if col not in header_columns]
-
-                if missing_input_columns:
-                    raise ValueError(f"Input columns not found in CSV header: {missing_input_columns}")
-                if missing_output_columns:
-                    raise ValueError(f"Expected output columns not found in CSV header: {missing_output_columns}")
-
-                # Get metadata columns (all columns not used for input or expected output)
-                metadata_columns = [
-                    col for col in header_columns if col not in input_columns and col not in expected_output_columns
-                ]
-
-                for row in rows:
-                    # Handle input data
-                    if len(input_columns) == 1:
-                        input_data = row[input_columns[0]]
-                    else:
-                        input_data = {col: row[col] for col in input_columns}
-
-                    # Handle expected output data
-                    if len(expected_output_columns) == 1:
-                        expected_output_data = row[expected_output_columns[0]]
-                    else:
-                        expected_output_data = {col: row[col] for col in expected_output_columns}
-
-                    # Handle metadata (all remaining columns)
-                    metadata = {col: row[col] for col in metadata_columns}
-
-                    data.append(
-                        {
-                            "input": input_data,
-                            "expected_output": expected_output_data,
-                            **metadata,
-                        }
-                    )
-        except FileNotFoundError as e:
-            raise DatasetFileError(f"CSV file not found: {filepath}") from e
-        except PermissionError as e:
-            raise DatasetFileError(f"Permission denied when reading CSV file: {filepath}") from e
-        except csv.Error as e:
-            raise DatasetFileError(f"Error parsing CSV file: {e}") from e
-        except Exception as e:
-            raise DatasetFileError(f"Unexpected error reading CSV file: {e}") from e
-
-        return cls(name=name, data=data, description=description)
-
-    def as_dataframe(self, multiindex: bool = True) -> "pd.DataFrame":
-        """
-        Convert the dataset to a pandas DataFrame for further analysis and manipulation.
-
-        Args:
-            multiindex (bool, optional): If True, expand 'input' and 'expected_output' dictionaries
-                into columns using a pandas MultiIndex. If False, keep them as columns containing
-                raw dictionaries.
-
-        Returns:
-            pd.DataFrame: DataFrame representation of the dataset.
-
-        Raises:
-            ImportError: If pandas is not installed.
-        """
-        try:
-            import pandas as pd
-        except ImportError:
-            raise ImportError(
-                "pandas is required to convert dataset to DataFrame. " "Please install it with `pip install pandas`"
-            )
-
-        if multiindex:
-            column_tuples = set()
-            data_rows = []
-            for record in self._data:
-                flat_record = {}
-
-                # Handle 'input' fields
-                input_data = record.get("input", {})
-                if isinstance(input_data, dict):
-                    for k, v in input_data.items():
-                        flat_record[("input", k)] = v
-                        column_tuples.add(("input", k))
-                else:
-                    flat_record[("input", "")] = input_data
-                    column_tuples.add(("input", ""))
-
-                # Handle 'expected_output' fields
-                expected_output = record.get("expected_output", {})
-                if isinstance(expected_output, dict):
-                    for k, v in expected_output.items():
-                        flat_record[("expected_output", k)] = v
-                        column_tuples.add(("expected_output", k))
-                else:
-                    flat_record[("expected_output", "")] = expected_output
-                    column_tuples.add(("expected_output", ""))
-
-                # Handle any other top-level fields
-                for k, v in record.items():
-                    if k not in ["input", "expected_output"]:
-                        flat_record[("metadata", k)] = v
-                        column_tuples.add(("metadata", k))
-                data_rows.append(flat_record)
-
-            # Convert column_tuples to a sorted list to maintain consistent column order
-            column_tuples = sorted(list(column_tuples))
-
-            # Build the DataFrame
-            records_list = []
-            for flat_record in data_rows:
-                row = [flat_record.get(col, None) for col in column_tuples]
-                records_list.append(row)
-
-            df = pd.DataFrame(records_list, columns=pd.MultiIndex.from_tuples(column_tuples))
-
-            return df
-
-        data = []
-        for record in self._data:
-            new_record = {}
-            input_data = record.get("input", {})
-            new_record["input"] = input_data
-            expected_output = record.get("expected_output", {})
-            new_record["expected_output"] = expected_output
-            # Copy other fields
-            for k, v in record.items():
-                if k not in ["input", "expected_output"]:
-                    new_record[k] = v
-            data.append(new_record)
-        return pd.DataFrame(data)
-
-    def __repr__(self) -> str:
-        """
-        Return a readable string representation of the Dataset object, including information such as
-        dataset name, size, structure, and synchronization status with Datadog.
-
-        Returns:
-            str: A human-friendly representation of the Dataset.
-        """
-        name = f"{Color.CYAN}{self.name}{Color.RESET}"
-        record_count = len(self._data)
-
-        # Get sample structure from first record
-        structure = []
-        if self._data:
-            sample = self._data[0]
-            # Input structure
-            input_data = sample.get("input", {})
-            if isinstance(input_data, dict):
-                input_desc = f"dict[{len(input_data)} keys]"
-            else:
-                input_desc = type(input_data).__name__
-            structure.append(f"input: {input_desc}")
-
-            # Expected output structure
-            expected = sample.get("expected_output", {})
-            if isinstance(expected, dict):
-                output_desc = f"dict[{len(expected)} keys]"
-            else:
-                output_desc = type(expected).__name__
-            structure.append(f"expected_output: {output_desc}")
-
-            # Metadata fields
-            metadata = {k: v for k, v in sample.items() if k not in ["input", "expected_output", "record_id"]}
-            if metadata:
-                structure.append(f"metadata: {len(metadata)} fields")
-
-        # Datadog sync status
-        if getattr(self, "_datadog_dataset_id", None):
-            dd_status = f"{Color.GREEN}✓ Synced{Color.RESET} (v{self._version})"
-            dd_url = f"\n  URL: {Color.BLUE}{BASE_URL}/llm/testing/datasets/{self._datadog_dataset_id}{Color.RESET}"
-        else:
-            dd_status = f"{Color.YELLOW}Local only{Color.RESET}"
-            dd_url = ""
-
-        # Build the representation
-        info = [
-            f"Dataset(name={name})",
-            f"  Records: {record_count:,}",
-            f"  Structure: {', '.join(structure)}",
-            f"  Datadog: {dd_status}",
-        ]
-
-        # Add description if present
-        if self.description:
-            desc_preview = (self.description[:47] + "...") if len(self.description) > 50 else self.description
-            info.insert(1, f"  Description: {desc_preview}")
-
-        # Add URL if dataset is synced
-        if dd_url:
-            info.append(dd_url)
-
-        return "\n".join(info)
-
+if TYPE_CHECKING:
+    import pandas as pd
 
 class Experiment:
     """
@@ -1725,374 +1056,3 @@ class ExperimentResults:
         return "\n".join(info)
 
 
-def _make_id() -> str:
-    """
-    Generate a unique identifier used for internal references in results or records.
-
-    Returns:
-        str: A random UUID as a hexadecimal string
-    """
-    return uuid.uuid4().hex
-
-
-def exp_http_request(method: str, url: str, body: Optional[bytes] = None) -> HTTPResponse:
-    """
-    Make an HTTP request to the Datadog LLM Observability/Experiments API.
-
-    Args:
-        method (str): HTTP method (e.g., "GET", "POST", "PATCH").
-        url (str): The full endpoint URL (path is appended to BASE_URL).
-        body (Optional[bytes]): Request body in bytes (for POST/PATCH).
-
-    Returns:
-        HTTPResponse: A wrapper object containing the response status code, headers, and body.
-
-    Raises:
-        ValueError: If the request fails with a 4xx or 5xx response from Datadog.
-    """
-    headers: Dict[str, str] = {
-        "DD-API-KEY": os.getenv(ENV_DD_API_KEY, ""),
-        "DD-APPLICATION-KEY": os.getenv(ENV_DD_APPLICATION_KEY, ""),
-        "Content-Type": "application/json",
-    }
-    full_url = BASE_URL + url
-    resp = http_request(method, full_url, headers=headers, body=body)
-    if resp.status_code == 403:
-        if not ENV_DD_SITE:
-            raise ValueError(
-                "DD_SITE may be incorrect. Please check your DD_SITE environment variable or pass it as an argument to init(...site=...)"
-            )
-        else:
-            raise ValueError(
-                "DD_API_KEY or DD_APPLICATION_KEY is incorrect. Please check your DD_API_KEY and DD_APPLICATION_KEY environment variables or pass them as arguments to init(...api_key=..., application_key=...)"
-            )
-    if resp.status_code >= 400:
-        try:
-            error_details = resp.json()
-            error_message = error_details.get("errors", [{}])[0].get("detail", resp.text())
-        except Exception:
-            error_message = resp.text()
-        raise ValueError(f"Request failed with status code {resp.status_code}: {error_message}")
-    return resp
-
-
-def task(func):
-    if func.__name__ == "task":
-        raise ValueError("Function name 'task' is reserved. Please use a different name for your task function.")
-
-    @wraps(func)
-    def wrapper(input: Dict[str, Union[str, Dict[str, Any]]], config: Optional[Dict[str, Any]] = None) -> Any:
-        # Call the original function with or without config
-        if "config" in inspect.signature(func).parameters:
-            return func(input, config)
-        return func(input)
-
-    # Enforce signature compliance
-    sig = inspect.signature(func)
-    params = sig.parameters
-    if "input" not in params:
-        raise TypeError("Task function must have an 'input' parameter.")
-    # Set attribute to indicate whether the function accepts config
-    wrapper._accepts_config = "config" in params
-    wrapper._is_task = True  # Set attribute to indicate decoration
-    return wrapper
-
-
-def evaluator(func):
-    @wraps(func)
-    def wrapper(
-        input: Union[str, Dict[str, Any]] = None,
-        output: Union[str, Dict[str, Any]] = None,
-        expected_output: Union[str, Dict[str, Any]] = None,
-    ) -> Any:
-        return func(input, output, expected_output)
-
-    # Enforce signature compliance
-    sig = inspect.signature(func)
-    params = sig.parameters
-    required_params = ["input", "output", "expected_output"]
-    if not all(param in params for param in required_params):
-        raise TypeError(f"Evaluator function must have parameters {required_params}.")
-    wrapper._is_evaluator = True  # Set attribute to indicate decoration
-    return wrapper
-
-
-def _print_progress_bar(iteration, total, prefix="", suffix="", decimals=1, length=50, fill="█"):
-    """
-    Print a simple progress bar to the console (commonly used for chunked uploads).
-
-    Args:
-        iteration (int): Current iteration or chunk number.
-        total (int): Total number of iterations or chunks.
-        prefix (str, optional): Optional prefix string. Defaults to "".
-        suffix (str, optional): Optional suffix string. Defaults to "".
-        decimals (int, optional): Number of decimal places to display in the percentage. Defaults to 1.
-        length (int, optional): Character length of the progress bar. Defaults to 50.
-        fill (str, optional): Character used to fill the progress bar. Defaults to "█".
-    """
-    percent = f"{100 * (iteration / float(total)):.{decimals}f}"
-    filled_length = int(length * iteration // total)
-    bar = fill * filled_length + "-" * (length - filled_length)
-    print(f"\r{prefix} |{bar}| {percent}% {suffix}", end="\r", flush=True)
-    if iteration == total:
-        print()
-
-
-class DatasetFileError(Exception):
-    """
-    Exception raised when there are errors reading or processing dataset files,
-    such as CSV or JSON errors or file permission issues.
-    """
-
-    pass
-
-
-class ExperimentTaskError(Exception):
-    """
-    Exception raised when a task fails during experiment execution.
-
-    Attributes:
-        message (str): Error message describing the failure.
-        row_idx (int): The index of the dataset record on which the task failed.
-        original_error (Exception): The original error object (if any).
-    """
-
-    def __init__(self, message: str, row_idx: int, original_error: Optional[Exception] = None) -> None:
-        self.row_idx = row_idx
-        self.original_error = original_error
-        super().__init__(message)
-
-
-class Color:
-    """
-    ANSI color codes for terminal output, which can be used to highlight or style text
-    in logs or console applications.
-    """
-
-    GREY = "\033[90m"
-    RED = "\033[91m"
-    GREEN = "\033[92m"
-    YELLOW = "\033[93m"
-    BLUE = "\033[94m"
-    MAGENTA = "\033[95m"
-    CYAN = "\033[96m"
-    WHITE = "\033[97m"
-    BOLD = "\033[1m"
-    DIM = "\033[2m"
-    RESET = "\033[0m"
-
-
-class Spinner:
-    """
-    Contains patterns for displaying animated spinners in the console,
-    used alongside progress or waiting states.
-    """
-
-    DOTS = ["⠋", "⠙", "⠹", "⠸", "⠼", "⠴", "⠦", "⠧", "⠇", "⠏"]
-    ARROW = ["▹▹▹▹▹", "▸▹▹▹▹", "▹▸▹▹▹", "▹▹▸▹▹", "▹▹▹▸▹", "▹▹▹▹▸"]
-
-
-class ProgressState(Enum):
-    """
-    Enumeration indicating overall progress states for console animations or logs.
-
-    Values:
-        RUNNING: The progress is currently in progress.
-        SUCCESS: The progress has completed successfully.
-        ERROR: The progress encountered an error.
-    """
-
-    RUNNING = "running"
-    SUCCESS = "success"
-    ERROR = "error"
-
-
-class ProgressReporter:
-    """
-    An enhanced progress reporter with color-coded output and optional "spinner"
-    animations, used for interactive console feedback during dataset processing.
-    """
-
-    def __init__(self, total: int, desc: str = "", width: Optional[int] = None) -> None:
-        """
-        Create a ProgressReporter to track completion and errors over a specified total.
-
-        Args:
-            total (int): The total number of items or iterations.
-            desc (str, optional): A short description of what is being tracked. Defaults to "".
-            width (int, optional): The width of the progress bar. If None, attempts to auto-detect terminal width.
-        """
-        self.total = total
-        self.current = 0
-        self.width: int = self._get_width(width)
-        self.start_time = time.time()
-        self.desc = desc
-        self.error_count = 0
-        self._last_len = 0
-        self._last_update = self.start_time
-        self._spinner_idx = 0
-        self._state = ProgressState.RUNNING
-
-    def _get_width(self, width: Optional[int]) -> int:
-        """
-        Determine the width of the progress bar, trying to respect terminal size if not specified.
-
-        Args:
-            width (int, optional): The desired progress bar width. Defaults to None.
-
-        Returns:
-            int: The width of the progress bar in characters.
-        """
-        if width is None:
-            try:
-                terminal_width = os.get_terminal_size().columns
-                return min(MAX_PROGRESS_BAR_WIDTH, terminal_width - 50)
-            except OSError:
-                return MAX_PROGRESS_BAR_WIDTH
-        return width
-
-    def update(self, advance: int = 1, error: bool = False) -> None:
-        """
-        Update progress with optional error tracking.
-
-        Args:
-            advance (int, optional): Number of items completed since last update. Defaults to 1.
-            error (bool, optional): Whether an error occurred. Defaults to False.
-        """
-        self.current += advance
-        if error:
-            self.error_count += 1
-            self._state = ProgressState.ERROR
-
-        # Throttle updates to max 15 per second
-        now = time.time()
-        if now - self._last_update < 0.066 and self.current < self.total:
-            return
-
-        self._last_update = now
-        self._spinner_idx = (self._spinner_idx + 1) % len(Spinner.DOTS)
-        self._print_progress()
-
-    def _format_time(self, seconds: float) -> str:
-        """
-        Return a user-friendly time format (e.g., 13.2s, 2m 4s, 1h 15m).
-
-        Args:
-            seconds (float): Elapsed time in seconds.
-
-        Returns:
-            str: A formatted time string.
-        """
-        if seconds < 60:
-            return f"{seconds:.1f}s"
-
-        total_minutes = int(seconds / 60)
-        remaining_seconds = seconds % 60
-
-        if total_minutes < 60:
-            return f"{total_minutes}m {int(remaining_seconds)}s"
-
-        total_hours = int(total_minutes / 60)
-        remaining_minutes = total_minutes % 60
-        return f"{total_hours}h {remaining_minutes}m"
-
-    def _format_speed(self) -> str:
-        """
-        Calculate and format processing speed.
-
-        Returns:
-            str: Formatted speed string.
-        """
-        elapsed = time.time() - self.start_time
-        if elapsed == 0:
-            return "0/s"
-        speed = self.current / elapsed
-        if speed >= 100:
-            return f"{speed:.0f}/s"
-        return f"{speed:.1f}/s"
-
-    def _get_gradient_color(self, progress: float) -> str:
-        """
-        Return a color based on progress and state.
-
-        Args:
-            progress (float): A float between 0 and 1 indicating progress fraction.
-
-        Returns:
-            str: An ANSI color code string.
-        """
-        if self._state == ProgressState.ERROR:
-            return Color.RED
-        elif progress >= 1:
-            return Color.GREEN
-        else:
-            return Color.BLUE
-
-    def _print_progress(self) -> None:
-        """
-        Print the current progress bar state to the console, including elapsed time, ETA, and error count.
-        """
-        elapsed = time.time() - self.start_time
-        progress = self.current / self.total if self.total > 0 else 1
-
-        if self.current == 0:
-            eta = 0
-        else:
-            speed = self.current / elapsed
-            remaining_items = self.total - self.current
-            eta = remaining_items / speed if speed > 0 else 0
-
-        color = self._get_gradient_color(progress)
-
-        # Build animated progress bar
-        filled_width = int(self.width * progress)
-        if progress < 1:
-            # Show animated gradient effect at progress point
-            bar = "█" * (filled_width - 1)
-            if filled_width > 0:
-                bar += "█"  # Pulse effect at progress point
-            bar += "░" * (self.width - filled_width)
-        else:
-            bar = "█" * self.width
-
-        # Format statistics
-        percent = f"{progress * 100:.1f}%"
-        counts = f"{self.current}/{self.total}"
-        speed = self._format_speed()
-        elapsed_str = self._format_time(elapsed)
-        eta_str = self._format_time(eta)
-
-        # Build error indicator if needed
-        if self.error_count > 0:
-            error_str = f" {Color.RED}({self.error_count} errors){Color.RESET}"
-        else:
-            error_str = ""
-
-        # Get spinner frame
-        spinner = Spinner.DOTS[self._spinner_idx] if progress < 1 else "✓"
-
-        # Construct status line with color
-        status = (
-            f"\r{spinner} {Color.BOLD}{self.desc}{Color.RESET} "
-            f"{color}|{bar}|{Color.RESET} "
-            f"{Color.BOLD}{percent}{Color.RESET} • "
-            f"{counts}{error_str} • "
-            f"{Color.DIM}{speed} • {elapsed_str}<{eta_str}{Color.RESET}"
-        )
-
-        # Clear previous line if needed
-        if self._last_len > len(status):
-            print(f"\r{' ' * self._last_len}", end="")
-
-        print(status, end="")
-        self._last_len = len(status)
-
-        # Print final status on completion
-        if self.current >= self.total:
-            print()
-            if self.error_count > 0:
-                error_rate = (self.error_count / self.total) * 100
-                print(f"{Color.RED}⚠️  Completed with {self.error_count} errors ({error_rate:.1f}%){Color.RESET}")
-                print(f"{Color.DIM}   Set .run(raise_errors=True) to halt and see full error traces{Color.RESET}")
-            else:
-                print(f"{Color.GREEN}✓ Completed successfully{Color.RESET}")
